requirements.txt
selenium~=4.21.0
pytest~=8.2.0
Faker~=26.0.0
allure-pytest==2.13.5
h11~=0.14.0
pip~=23.2.1
attrs~=23.2.0
typing_extensions~=4.12.2
trio~=0.25.1
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
astor~=0.8.1
idna~=3.7
python-dateutil~=2.9.0.post0
pluggy~=1.5.0
iniconfig~=2.0.0
certifi~=2024.2.2
urllib3~=2.2.1
six~=1.16.0
wsproto~=1.2.0
python-dotenv~=1.0.1